update offer section
