+++ src/App.jsx
@@ -3,7 +3,7 @@ import "./styles/config/main.scss";
 import Header from "./components/header";
 import Welcome from "./components/welcome";
 import About from "./components/about";
-import Specialization from "./components/specialization";
+import Offer from "./components/offer";
 import Services from "./components/services";
 import Contact from "./components/contact";
 
@@ -13,7 +13,7 @@ function App() {
       <Header />
       <Welcome />
       <About />
-      <Specialization />
+      <Offer />
       <Services />
       <Contact />
     </div>
