+++ src/styles/config/main.scss
@@ -10,7 +10,7 @@
 @import '../header.scss';
 @import '../welcome.scss';
 @import '../about.scss';
-@import '../specialization.scss';
+@import '../offer.scss';
 @import '../services.scss';
 @import '../contact.scss';
