--- src/styles/specialization.scss
@@ -1,88 +1,0 @@
-.spcialization-section {
-  background-color: $c-primary;
-  padding:30px 20px;
-  text-align: center;
-}
-
-.specialization-container {
-  padding: 20px 40px;
-  h3 {
-    color: $c-theme;
-    padding-bottom: 20px;
-  }
-}
-
-.specialization-picture {
-  position: relative;
-  color: $c-white;
-  background-image: url('../assets/relax.jpg');
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  box-sizing: border-box;
-  display: flex;
-  justify-content:center;
-  align-items: center;
-
-  &::before {
-    content: '';
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    background: rgba(0, 0, 0, 0.582); // Czarna nakładka z 50% nieprzezroczystości
-    z-index: 1;
-  }
-
-  > * {
-    position: relative;
-    z-index: 2;
-  }
-}
-
-
-@include for-tablet {
-  .spcialization-section {
-    padding: 30px 20px 50px 20px;
-  }
-
-  .specialization-container {
-    padding: 20px 80px;
-    h3 {
-     font-size: $header-size-semi-large;
-    }
-  }
-
-  .specialization-picture {
-    height: 160px;
-    p{
-      font-size: $header-size-medium;
-    }
-  }
-}
-
-
-@include for-desktop {
-  .spcialization-section {
-    padding: 40px 30px 60px 30px;
-  }
-
-  .specialization-container {
-    padding: 10px 200px 30px 200px;
-    h3 {
-     font-size: $header-size-semi-large;
-    }
-    p{
-      font-size: $font-size-semi-large;
-    }
-  }
-
-  .specialization-picture {
-    height: 160px;
-
-    p{
-      font-size: $header-size-medium;
-    }
-  }
-}
